Moved tests back to mock_worker.
